requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.7.0

# Google API dependencies
google-auth>=2.0.0
google-auth-oauthlib>=0.8.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# Firebase dependencies
firebase-admin>=6.2.0

# NLP and ML dependencies
spacy>=3.5.0
transformers>=4.20.0
torch>=1.12.0

# Web scraping and processing
beautifulsoup4>=4.11.0
requests>=2.28.0

# Google Generative AI
google-generativeai>=0.3.0

# Web framework
flask>=2.3.0
flask-cors>=4.0.0

# Additional utilities
python-dotenv>=0.19.0
